- added additional conditional for use_cached_merge_only to prevent deletion of merged files if this is set. this ensures that use_cached_merge_only properly overrides all other collate settings

+++ source/setup/setup.py
@@ -26,7 +26,7 @@ def tidy_existing_file_tree(root_dir):
 
     # Merged files. Orphans get left here if deleted from e.g. zotero or
     # the raw cache folder.
-    if config.merge_all == 1:
+    if config.merge_all == 1 and config.use_cached_merge_only == 0:
         if os.path.exists(config.cache_dir + '/processed/merged'):
             shutil.rmtree(config.cache_dir + '/processed/merged')
